tile and limits errors
