+++ src/vis/vis.js
@@ -18,6 +18,7 @@ var VisModel = Backbone.Model.extend({
   defaults: {
     loading: false,
     showEmptyInfowindowFields: false,
+    showErrors: true,
     state: STATE_INIT
   },
