+++ themes/scss/map/_limits.scss
@@ -33,6 +33,9 @@ $linkColor: #139BFC;
     }
   }
 }
+.CDB-Limits--short.is-active {
+  width: 260px;
+}
 
 .CDB-Limits-button {
   display: flex;
